1. Moved player file parsing into `FileManager`. 2. Removed some fields from the `MovementData` class and constructor.

+++ Shroomworld/Shroomworld/Components/ReadonlyMovementData.cs
@@ -8,7 +8,7 @@ namespace Shroomworld
 {
 	internal class ReadonlyMovementData
 	{
-		public int MovementForce => _movementForce;
+		public float MovementForce => _movementForce;
 		public int Mass => _mass;
